requirements.txt
aiohttp==3.8.4
docstring-parser==0.15
omegaconf
joblib==1.2.0
hydra-core==1.3.2
ipykernel==6.22.0
jsonargparse==4.20.1
netcdf4==1.6.3
opencv-python==4.7.0.72
pip-chill==1.0.3
scikit-image==0.20.0
seaborn==0.12.2
streamlit==1.21.0
tqdm==4.65.0
tensorboard==2.12.2
tensorboardx==2.6
torchaudio==2.0.1+cu118
torchinfo==1.7.2
torchsummary==1.5.1
graphviz==0.20.1
torchviz==0.0.2
typeshed-client==2.2.0
scikit-learn==1.2.2
shapely==2.0.1
pywin32==306
PyYAML==6.0
Pillow
pandas
numpy